# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/inventory/movements": {
      "get": {
        "operationId": "InventoryController_listMovements",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List inventory movements"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "inventory"
        ]
      }
    },
    "/inventory/reservations": {
      "get": {
        "operationId": "InventoryController_listReservations",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List inventory reservations"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "inventory"
        ]
      }
    },
    "/inventory/alerts/low-stock": {
      "get": {
        "operationId": "InventoryController_listLowStockAlerts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List low-stock inventory alerts"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "inventory"
        ]
      }
    },
    "/inventory/variants/{variantId}/withdrawal-priority": {
      "get": {
        "operationId": "InventoryController_listVariantWithdrawalPriority",
        "parameters": [
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List warehouse withdrawal priority for variant"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "inventory"
        ]
      }
    },
    "/inventory/variants/{variantId}/adjustments": {
      "post": {
        "operationId": "InventoryController_adjustVariant",
        "parameters": [
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustInventoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Adjust variant stock and create movement record"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "inventory"
        ]
      }
    },
    "/inventory/variants/{variantId}/threshold": {
      "put": {
        "operationId": "InventoryController_updateLowStockThreshold",
        "parameters": [
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLowStockThresholdDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update low-stock threshold for a variant"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "inventory"
        ]
      }
    },
...
  },
  "info": {
    "title": "General Ecommerce API",
    "description": "General Ecommerce backend APIs",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AdjustInventoryDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateLowStockThresholdDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}