# ecommerce-core-api/src/inventory/dto/adjust-inventory.dto.ts
import { IsInt, IsOptional, IsString, IsUUID, MaxLength, NotEquals } from 'class-validator';

export class AdjustInventoryDto {
  @IsUUID('4')
  warehouseId!: string;

  @IsInt()
  @NotEquals(0)
  quantityDelta!: number;

  @IsOptional()
  @IsString()
  @MaxLength(500)
  note?: string;

  @IsOptional()
  @IsString()
  @MaxLength(100)
  idempotencyKey?: string;
}